Insurance_app_b/backend/api/serializers.py
from rest_framework import serializers
from .models import Vehicle, UserProfile

class VehicleSerializer(serializers.ModelSerializer):
    class Meta:
        model = Vehicle
        fields = '__all__'

class UserProfileSerializer(serializers.ModelSerializer):
    class Meta:
        model = UserProfile
        fields = '__all__'
