Insurance_app_b/backend/requirements.txt
asgiref==3.8.1
Django==4.2.17
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
mysqlclient==2.2.6
PyJWT==2.10.1
PyMySQL==1.1.1
python-dotenv==1.0.1
sqlparse==0.5.3
typing_extensions==4.12.2
